backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
scikit-learn
tensorflow
pandas
numpy
pydantic
requests